refactor: modularize analysis logic into package (Issue #34)
