+++ src/mcp_repo_onboarding/analysis/__init__.py
@@ -1,0 +1,20 @@
+from .core import analyze_repo
+from .extractors import (
+    detect_workflow_python_version,
+    extract_makefile_commands,
+    extract_shell_scripts,
+    extract_tox_commands,
+)
+from .scanning import scan_repo_files
+from .structs import SAFETY_IGNORES, IgnoreMatcher
+
+__all__ = [
+    "analyze_repo",
+    "extract_makefile_commands",
+    "extract_shell_scripts",
+    "extract_tox_commands",
+    "detect_workflow_python_version",
+    "IgnoreMatcher",
+    "SAFETY_IGNORES",
+    "scan_repo_files",
+]
